requirements.txt
pika
pymongo
sco-datastore
